locust/requirements.txt
locust==2.37.7
requests>=2.32.2
python-dotenv==1.0.1
